requirements.txt
asgiref~=3.3.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography~=3.2.1
distlib~=0.3.2
Django~=2.2.24
django-allauth~=0.45.0
django-rest-auth==0.9.5
djangorestframework~=3.11.2
djoser~=2.0.5
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe~=1.1.1
oauthlib~=3.1.0
packaging~=20.9
psycopg2~=2.7.7
pycparser==2.20
Pygments~=2.9.0
PyJWT~=1.7.1
pyparsing==2.4.7
pyrsistent~=0.17.3
pytz==2021.1
PyYAML~=5.3.1
requests==2.25.1
simplejson~=3.17.3
six==1.16.0
soupsieve~=2.1
sqlparse==0.4.1
tornado==6.1
uritemplate==3.0.1
urllib3==1.26.6
virtualenv~=20.6.0

pbr~=5.6.0
setuptools~=50.3.2
pip~=20.3.4
stevedore~=1.32.0
platformdirs~=2.0.2
virtualenvwrapper~=4.8.4
zipp~=1.2.0
python-dotenv~=0.19.0
pytest~=6.1.2
py~=1.10.0
toml~=0.10.2
pluggy~=0.13.1
iniconfig~=1.1.1
atomicwrites~=1.4.0
pathlib2~=2.3.6